+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - `nullifzero`
   - `snowflake_cortex_sentiment`
 - Added `Catalog` class to manage snowflake objects. It can be accessed via `Session.catalog`.
+- Added support for `DataFrameWriter.insert_into/insertInto`.
 
 #### Improvements